skills/exa-grounding/requirements.txt
exa-py
vecs
python-dotenv
